content/xul/content/test/test_bug398289.html

<!DOCTYPE html>
<html class="reftest-wait" style="height: 100%">
<head>
  <title>Test for bug 398289</title>
  <script type="application/javascript" src="chrome://mochikit/content/MochiKit/packed.js"></script>
  <script type="application/javascript" src="chrome://mochikit/content/tests/SimpleTest/SimpleTest.js"></script>
  <link rel="stylesheet" type="text/css" href="chrome://mochikit/content/tests/SimpleTest/test.css" />
</head>
<body style="height: 100%">
  <iframe id="test" src="chrome://mochikit/content/chrome/content/xul/content/test/398289-resource.xul" width="100%" height="100%">
  </iframe>

  <script class="testbody" type="text/javascript">
    SimpleTest.waitForExplicitFinish();
  
    window.frames[0].document.getElementById("test").selectedIndex = 0;
    window.frames[0].document.getElementById("test").selectedIndex = 1;
    document.getElementById("test").onload = onLoad;
    window.frames[0].location.reload();
    
    function onLoad() {
      document.documentElement.className = "";
      is(window.frames[0].document.getElementById("test").selectedIndex, 1, 
          "persistent attribute in tab box broken");
          
      SimpleTest.finish();
    }
  </script>
</body>
</html>
